+++ src/onegov.people/onegov/people/models/person.py
@@ -1,12 +1,15 @@
 from onegov.core.orm import Base
+from onegov.core.orm.mixins import ContentMixin
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
 from onegov.search import ORMSearchable
-from sqlalchemy import Column, Text
+from sqlalchemy import Column
+from sqlalchemy import Text
 from uuid import uuid4
 
 
-class Person(Base, TimestampMixin, ORMSearchable):
+class Person(Base, ContentMixin, TimestampMixin, ORMSearchable):
+    """ A person. """
 
     __tablename__ = 'people'
 
@@ -30,10 +33,17 @@ class Person(Base, TimestampMixin, ORMSearchable):
 
     @property
     def title(self):
+        """ Returns the Estern-ordered name. """
+
         return self.last_name + " " + self.first_name
 
     @property
     def spoken_title(self):
+        """ Returns the Western-ordered name. Includes the salutation if
+        available.
+
+        """
+
         if self.salutation:
             parts = self.salutation, self.first_name, self.last_name
         else:
@@ -44,22 +54,32 @@ class Person(Base, TimestampMixin, ORMSearchable):
     #: the unique id, part of the url
     id = Column(UUID, primary_key=True, default=uuid4)
 
+    #: the salutation used for the person
     salutation = Column(Text, nullable=True)
 
+    #: the first name of the person
     first_name = Column(Text, nullable=False)
 
+    #: the last name of the person
     last_name = Column(Text, nullable=False)
 
+    #: the function of the person
     function = Column(Text, nullable=True)
 
+    #: an URL leading to a picture of the person
     picture_url = Column(Text, nullable=True)
 
+    #: the email of the person
     email = Column(Text, nullable=True)
 
+    #: the phone number of the person
     phone = Column(Text, nullable=True)
 
+    #: the website related to the person
     website = Column(Text, nullable=True)
 
+    #: the address of the person
     address = Column(Text, nullable=True)
 
+    #: some remarks about the person
     notes = Column(Text, nullable=True)